fix(app): provider toast fires for first provider + modal closes after API key save + move defaults to app/defaults/
